Move palettes into their own directory
